requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
bcrypt
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
databases==0.9.0
dnspython==2.6.1
email-validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fpdf2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
jinja2==3.1.4
kombu==5.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.24.4
opencv-python==4.9.0.80
orjson==3.10.3
packaging==24.0
passlib==1.7.4
pillow==10.3.0
prompt-toolkit==3.0.45
psycopg2-binary==2.9.9
pydantic==2.7.2
pydantic-core==2.18.3
pygments==2.18.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.4
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.12.3
typing-extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
pandas